+++ src/controllers/Products.spec.js
@@ -120,31 +120,6 @@ describe('Products', () => {
             expect(resMock.sendStatus).toBeCalledWith(200);
         });
     });
-    describe('productsProductIdsGetOld (deprecated)', () => {
-        it('calls productsProductIdsGet method from service', async () => {
-            const resMock = generateResponseMock();
-            const reqMock = generateRequestMock();
-
-            service.productsProductIdsGet.mockResolvedValue({
-                products: testProducts,
-                total: 2,
-                hasNext: false
-            });
-            reqMock.query = {
-                lang: testLang
-            };
-            reqMock.params = {
-                productIds: testProductsParameter
-            };
-
-            await controller.productsProductIdsGet(reqMock, resMock);
-
-            expect(service.productsProductIdsGet.mock.calls.length).toBe(1);
-            expect(service.productsProductIdsGet.mock.calls[0][0]).toEqual(testProducts);
-            expect(service.productsProductIdsGet.mock.calls[0][1]).toEqual(testLang);
-            expect(writer.writeJson).toHaveBeenCalledTimes(1);
-        });
-    });
     describe('productsProductIdsHead', () => {
         it('fills the res with a 200 code', async () => {
             const resMock = generateResponseMock();